[cmds] Further tricky fixes for change in wait4/waitpid operation from #2653

+++ elkscmd/inet/ftpd.c
@@ -310,7 +310,7 @@ void sigalrm(int sig)
 void sigchild(int sig)
 {
 	signal(SIGCHLD, sigchild);
-	while (waitpid(-1, NULL, WNOHANG) > 0)
+	while ((int)waitpid(-1, NULL, WNOHANG) > 0)
 		continue;
 }
