+++ elkscmd/inet/telnetd.c
@@ -179,7 +179,7 @@ void
 sigchild(int sig)
 {
     signal(SIGCHLD, sigchild);
-    while (waitpid(-1, NULL, WNOHANG) > 0)
+    while ((int)waitpid(-1, NULL, WNOHANG) > 0)
         continue;
 }
